requirements.txt
streamlit==1.35.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit-authenticator==0.3.2
pandas==2.2.2